don't throw warnings if there is -l on the shebang
There were warnings about undefined variables when the program being debugged had the -l switch in its shebang line. This was due to the fact that -l changes $\, which added an additional \n to the serialized data that was written to the socket in the backend. The client then recieved output with two newlines, and had trouble digesting that, resulting in empty data about the basic debugging informaation like the current line (and possibly later other info).

This fixes RT 29956 (https://rt.cpan.org/Public/Bug/Display.html?id=29956) and closes #14.